move identity_key as a method of NetworkConfig
